% Generated by roxygen2 (4.0.1): do not edit by hand
\name{inverse.convex}
\alias{inverse.convex}
\title{Fits the Inverse Convex Estimator}
\usage{
  inverse.convex(x, max.it = 500, tol 10^-4, plot = FALSE, display = FALSE)
}
\arguments{
  \item{x}{data set to be fit with inverse convex estimator}
  \item{max.it}{maximum iterations of algorithm}
  \item{tol}{tolerance required for convergence}
  \item{plot}{should the estimated density be plotted?}
  \item{display}{should information about the convergence of the algorithm be displayed?}
}
\description{

Fits an estimate such it is restrained to being inverse convex, i.e. f(x) = 1/g(x) where g(x) is convex.
This fit is actually a local maximum, not a global: the likelihood function is unbounded. Much like
the gaussian mixture, although the likelihood is unbounded around a degenerate distribution,
a very useful estimator can be obtained from a local non-degenerate distribution. Unlike the gaussian
mixture, this local maximum appears unique. Furthermore, the domain of convergence to degenerate
mode (emprically) shrinks very quickly, such that this algorithm avoids that mode with very high probability 
as long there is at least 30 unique points in the dataset. 
  
}
\details{

  Returns an "ICfit" object, which contains "x", the location of the active points
  and "dens", the estimated density at the active points. These are the sufficient
  statistics describing an inverse convex fit. 


  Uses a sequential quadratic programming algorithm, paired with a univariate 
  optimization step. See Semi- and Non- Parametric Methods for Interval Censored Data
  with Shape Constraints
  (note: this algorithm cannot be applied to 
  interval censored data...yet)
}
\examples{
  fit = inverse.convex(rnorm(500) )
  # Fits an inverse convex estimator to a sample
  
  plot(fit)  							
  # Plots the estimated CDF
  
  qIC(0.5, fit)  						
  # Estimates the median
}
\author{
  Clifford Anderson-Bergman
}